BINEXP-82
* login failure message
* refactored error handling in forms
