+++ src/components/login/loginComponent.tsx
@@ -30,10 +30,8 @@ const LoginComponent = () => {
         const err = e.response?.data;
         if (err) {
             const errorMap = {
-                username: err?.username?.slice(0, 1),
-                password: err?.detail || err?.password?.slice(0, 1)
+                password: 'Niepoprawny login lub hasło.'
             };
-            console.log(errorMap);
             setErrors(errorMap);
         } else {
             setErrors({});
