+++ src/components/login/loginForm.tsx
@@ -5,6 +5,7 @@ import FormInputField from '../shared/form/formInputField';
 import { formTexts } from '../shared/form/formTexts';
 import SubmitButton from '../shared/form/submitButton';
 import FormHeaderComponent from '../shared/form/formHeader';
+import { handleBackendErrors } from '../shared/form/backendErrorHandler';
 
 interface LoginFormProps {
     loading: boolean;
@@ -20,20 +21,7 @@ const LoginForm = ({ loading, backendErrors, onSubmit }: LoginFormProps) => {
         formState: { errors }
     } = useForm<LoginData>({ mode: 'all' });
 
-    const handleBackendErrors = (field: string): FieldError | undefined => {
-        const message = backendErrors[field];
-        if (message === undefined) {
-            return undefined;
-        }
-
-        const err: FieldError = {
-            type: 'server',
-            message: message
-        };
-
-        setError(field as FieldName<LoginData>, err);
-        backendErrors[field] = undefined;
-    };
+    const handleErrors = (field: string) => handleBackendErrors<LoginData>(field, backendErrors, setError);
 
     return (
         <div className="flex items-center justify-center">
@@ -48,7 +36,7 @@ const LoginForm = ({ loading, backendErrors, onSubmit }: LoginFormProps) => {
                 <form className="mt-8 space-y-2" onSubmit={handleSubmit(onSubmit)} method="POST">
                     <FormInputField
                         placeholder={'Nazwa użytkownika'}
-                        errors={errors?.username || handleBackendErrors('username')}
+                        errors={errors?.username}
                         registerOpts={register('username', {
                             required: formTexts.requiredField
                         })}
@@ -58,7 +46,7 @@ const LoginForm = ({ loading, backendErrors, onSubmit }: LoginFormProps) => {
                     <FormInputField
                         type={'password'}
                         placeholder={'Hasło'}
-                        errors={errors?.password || handleBackendErrors('password')}
+                        errors={errors?.password || handleErrors('password')}
                         registerOpts={register('password', {
                             required: formTexts.requiredField
                         })}
