+++ src/components/register/registerForm.tsx
@@ -5,6 +5,8 @@ import { formTexts } from '../shared/form/formTexts';
 import FormInputField from '../shared/form/formInputField';
 import FormHeaderComponent from '../shared/form/formHeader';
 import SubmitButton from '../shared/form/submitButton';
+import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { LoginData } from '../../models/loginData';
 
 interface RegisterFormProps {
     loading: boolean;
@@ -21,20 +23,7 @@ const RegisterForm = ({ loading, backendErrors, onSubmit }: RegisterFormProps) =
         formState: { errors }
     } = useForm<RegisterData>({ mode: 'all' });
 
-    const handleBackendErrors = (field: string): FieldError | undefined => {
-        const message = backendErrors[field];
-        if (message === undefined) {
-            return undefined;
-        }
-
-        const err: FieldError = {
-            type: 'server',
-            message: message
-        };
-
-        setError(field as FieldName<RegisterData>, err);
-        backendErrors[field] = undefined;
-    };
+    const handleErrors = (field: string) => handleBackendErrors<RegisterData>(field, backendErrors, setError);
 
     const password = useRef({});
     password.current = watch('password', '');
@@ -52,7 +41,7 @@ const RegisterForm = ({ loading, backendErrors, onSubmit }: RegisterFormProps) =
                 <form className="mt-8 space-y-2" onSubmit={handleSubmit(onSubmit)} method="POST">
                     <FormInputField
                         placeholder={'Nazwa użytkownika'}
-                        errors={errors?.username || handleBackendErrors('username')}
+                        errors={errors?.username || handleErrors('username')}
                         registerOpts={register('username', {
                             required: formTexts.requiredField,
                             minLength: {
@@ -65,7 +54,7 @@ const RegisterForm = ({ loading, backendErrors, onSubmit }: RegisterFormProps) =
 
                     <FormInputField
                         placeholder={'Email'}
-                        errors={errors?.email || handleBackendErrors('email')}
+                        errors={errors?.email || handleErrors('email')}
                         registerOpts={register('email', {
                             required: formTexts.requiredField,
                             pattern: {
