+++ src/components/settings/deleteButton.tsx
@@ -14,6 +14,8 @@ import { useHistory } from 'react-router-dom';
 import { useAppDispatch } from '../../context/hooks';
 import { useToasts } from 'react-toast-notifications';
 import { DeleteAccountData } from '../../models/deleteAccountData';
+import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { RegisterData } from '../../models/registerData';
 
 interface DeleteButtonProps {}
 
@@ -42,20 +44,7 @@ const DeleteAccountButton = (props: PropsWithChildren<DeleteButtonProps>) => {
         setModalIsOpen(false);
     };
 
-    const handleBackendErrors = (field: string): FieldError | undefined => {
-        const message = backendErrors[field];
-        if (message === undefined) {
-            return undefined;
-        }
-
-        const err: FieldError = {
-            type: 'server',
-            message: message
-        };
-
-        setError(field as FieldName<DeleteAccountData>, err);
-        backendErrors[field] = undefined;
-    };
+    const handleErrors = (field: string) => handleBackendErrors<DeleteAccountData>(field, backendErrors, setError);
 
     const deletedCallback = () => {
         setLoading(false);
@@ -131,7 +120,7 @@ const DeleteAccountButton = (props: PropsWithChildren<DeleteButtonProps>) => {
                                 <FormInputField
                                     type={'password'}
                                     placeholder={'Aktualne hasło'}
-                                    errors={errors?.password || handleBackendErrors('password')}
+                                    errors={errors?.password || handleErrors('password')}
                                     registerOpts={register('password', {
                                         required: formTexts.requiredField
                                     })}
