+++ src/components/settings/settingsForm.tsx
@@ -8,6 +8,8 @@ import { SettingsData } from '../../models/settingsData';
 import { useAppSelector } from '../../context/hooks';
 import { currentUserData } from '../../context/userReducer';
 import DeleteAccountButton from './deleteButton';
+import { handleBackendErrors } from '../shared/form/backendErrorHandler';
+import { DeleteAccountData } from '../../models/deleteAccountData';
 
 interface RegisterFormProps {
     loading: boolean;
@@ -25,20 +27,7 @@ const SettingsForm = ({ loading, backendErrors, onSubmit }: RegisterFormProps) =
         formState: { errors }
     } = useForm<SettingsData>({ mode: 'all' });
 
-    const handleBackendErrors = (field: string): FieldError | undefined => {
-        const message = backendErrors[field];
-        if (message === undefined) {
-            return undefined;
-        }
-
-        const err: FieldError = {
-            type: 'server',
-            message: message
-        };
-
-        setError(field as FieldName<SettingsData>, err);
-        backendErrors[field] = undefined;
-    };
+    const handleErrors = (field: string) => handleBackendErrors<SettingsData>(field, backendErrors, setError);
 
     const newPassword = useRef({});
     const newEmail = useRef({});
@@ -60,7 +49,7 @@ const SettingsForm = ({ loading, backendErrors, onSubmit }: RegisterFormProps) =
                     <div className="col-span-2">
                         <FormInputField
                             placeholder={'Nowy email'}
-                            errors={errors?.email || handleBackendErrors('email')}
+                            errors={errors?.email || handleErrors('email')}
                             registerOpts={register('email', {
                                 pattern: {
                                     value: /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i,
@@ -103,7 +92,7 @@ const SettingsForm = ({ loading, backendErrors, onSubmit }: RegisterFormProps) =
                         <FormInputField
                             type={'password'}
                             placeholder={'Aktualne hasło'}
-                            errors={errors?.oldPassword || handleBackendErrors('oldPassword')}
+                            errors={errors?.oldPassword || handleErrors('oldPassword')}
                             registerOpts={register('oldPassword', {
                                 required: formTexts.requiredField,
                                 validate: (value: string) => {
