+++ src/components/shared/form/backendErrorHandler.ts
@@ -1,0 +1,20 @@
+import { FieldError, FieldName, FieldPath, UseFormSetError } from 'react-hook-form';
+
+export const handleBackendErrors = <T>(
+    field: string,
+    backendErrors: Record<string, string | undefined>,
+    setError: UseFormSetError<T>
+): FieldError | undefined => {
+    const message = backendErrors[field];
+    if (message === undefined) {
+        return undefined;
+    }
+
+    const err: FieldError = {
+        type: 'server',
+        message: message
+    };
+
+    setError(field as FieldPath<T>, err);
+    backendErrors[field] = undefined;
+};
