+++ src/components/tasks/single/solveForm.tsx
@@ -4,6 +4,8 @@ import { SolveData } from '../../../models/solveData';
 import { formTexts } from '../../shared/form/formTexts';
 import FormInputField from '../../shared/form/formInputField';
 import SubmitButton from '../../shared/form/submitButton';
+import { handleBackendErrors } from '../../shared/form/backendErrorHandler';
+import { SettingsData } from '../../../models/settingsData';
 
 interface SolveFormProps {
     loading: boolean;
@@ -19,20 +21,7 @@ const SolveForm = ({ loading, backendErrors, onSubmit }: SolveFormProps) => {
         formState: { errors }
     } = useForm<SolveData>({ mode: 'onSubmit' });
 
-    const handleBackendErrors = (field: string): FieldError | undefined => {
-        const message = backendErrors[field];
-        if (message === undefined) {
-            return undefined;
-        }
-
-        const err: FieldError = {
-            type: 'server',
-            message: message
-        };
-
-        setError(field as FieldName<SolveData>, err);
-        backendErrors[field] = undefined;
-    };
+    const handleErrors = (field: string) => handleBackendErrors<SolveData>(field, backendErrors, setError);
 
     return (
         <div className="w-full mt-6">
@@ -40,7 +29,7 @@ const SolveForm = ({ loading, backendErrors, onSubmit }: SolveFormProps) => {
                 <div className="mx-auto py-1 flex gap-4">
                     <FormInputField
                         placeholder={'flAGH{example_flag}'}
-                        errors={errors?.flag || handleBackendErrors('flag')}
+                        errors={errors?.flag || handleErrors('flag')}
                         registerOpts={register('flag', {
                             required: formTexts.requiredField
                         })}
